fix build for node version 10 (dependency upgrade -> "nan": "^2.10.0")
I just found this similar issue and fix in another package here: https://github.com/noble/node-bluetooth-hci-socket/issues/84#issuecomment-391951781

The same seems to work here, to make it build with Node v10. Hope it does not break anything with earlier versions, as long as they update the NaN dependency to ^2.10.